+++ src/constants.ts
@@ -1,4 +1,11 @@
 export const MODULE_NAME = "fast-flip";
-export const MIRROR_HORIZONTAL_HOT_KEY = `${MODULE_NAME}.mirror-horizontal-hotkey`;
-export const MIRROR_VERTICAL_HOT_KEY = `${MODULE_NAME}.mirror-vertical-hotkey`;
-export const TOGGLE_AFK_HOT_KEY = `${MODULE_NAME}.toggle-afk-hotkey`;
+export const MIRROR_HORIZONTAL_HOTKEY = `${MODULE_NAME}.mirror-horizontal-hotkey`;
+export const MIRROR_HORIZONTAL_HINT = `${MODULE_NAME}.mirror-horizontal-hint`;
+export const MIRROR_HORIZONTAL_BUTTON = `${MODULE_NAME}.mirror-horizontal-button`;
+export const MIRROR_VERTICAL_HOTKEY = `${MODULE_NAME}.mirror-vertical-hotkey`;
+export const MIRROR_VERTICAL_HINT = `${MODULE_NAME}.mirror-vertical-hint`;
+export const MIRROR_VERTICAL_BUTTON = `${MODULE_NAME}.mirror-vertical-button`;
+export const TOGGLE_AFK_HOTKEY = `${MODULE_NAME}.toggle-afk-hotkey`;
+export const TOGGLE_AFK_HINT = `${MODULE_NAME}.toggle-afk-hint`;
+export const TOGGLE_AFK_BUTTON = `${MODULE_NAME}.toggle-afk-button`;
+