+++ src/module.ts
@@ -1,6 +1,7 @@
-import { MODULE_NAME, MIRROR_HORIZONTAL_HOT_KEY, MIRROR_VERTICAL_HOT_KEY, TOGGLE_AFK_HOT_KEY } from "./constants";
+import * as constants from "./constants";
 import { TokenManager, TokenMirror } from "./managers/TokenManager";
 import { TileManager, TileMirror } from "./managers/TileManager";
+import { TokenHUDManager } from "managers/TokenHUDManager";
 
 Hooks.once("init", () => {
     if (game instanceof Game) {
@@ -9,20 +10,25 @@ Hooks.once("init", () => {
 });
 
 class FastFlipModule {
+    readonly #tokenHUDManager: TokenHUDManager;
     readonly #tokenManager: TokenManager;
     readonly #tileManager: TileManager;
+    readonly #game: Game;
 
     constructor(game: Game) {
+        this.#game = game;
+        this.#tokenHUDManager = new TokenHUDManager(game);
         this.#tokenManager = new TokenManager(game);
         this.#tileManager = new TileManager(game);
 
-        this.#registerKeybindings(game.keybindings);
+        this.#registerKeybindings();
+        this.#registerHUDButtons();
     }
 
-    #registerKeybindings(keybindings: ClientKeybindings) {
-        keybindings.register(MODULE_NAME, "horizontalFlip", {
-            name: MIRROR_HORIZONTAL_HOT_KEY,
-            hint: "Horizontally mirrors the selected tile or token",
+    #registerKeybindings() {
+        this.#game.keybindings.register(constants.MODULE_NAME, "horizontalFlip", {
+            name: constants.MIRROR_HORIZONTAL_HOTKEY,
+            hint: this.#game.i18n.localize(constants.MIRROR_HORIZONTAL_HINT),
             editable: [
                 { key: "KeyF" },
             ],
@@ -34,9 +40,9 @@ class FastFlipModule {
             repeat: false,
         });
 
-        keybindings.register(MODULE_NAME, "verticalFlip", {
-            name: MIRROR_VERTICAL_HOT_KEY,
-            hint: "Vertically mirrors the selected tile or token",
+        this.#game.keybindings.register(constants.MODULE_NAME, "verticalFlip", {
+            name: constants.MIRROR_VERTICAL_HOTKEY,
+            hint: this.#game.i18n.localize(constants.MIRROR_VERTICAL_HINT),
             editable: [
                 { key: "KeyF", modifiers: ["SHIFT"] },
             ],
@@ -48,9 +54,9 @@ class FastFlipModule {
             repeat: false,
         });
 
-        keybindings.register(MODULE_NAME, TOGGLE_AFK_HOT_KEY, {
-            name: MIRROR_VERTICAL_HOT_KEY,
-            hint: "Marks the selected tokens as AFK",
+        this.#game.keybindings.register(constants.MODULE_NAME, constants.TOGGLE_AFK_HOTKEY, {
+            name: constants.TOGGLE_AFK_HOTKEY,
+            hint: this.#game.i18n.localize(constants.TOGGLE_AFK_HINT),
             editable: [
                 { key: "KeyK", modifiers: ["SHIFT"] },
             ],
@@ -60,6 +66,29 @@ class FastFlipModule {
             restrictied: false,
             reservedModifiers: [],
             repeat: false,
+        });
+    }
+
+    #registerHUDButtons() {
+        this.#tokenHUDManager.registerButton(`${constants.MODULE_NAME}.mirror-horizontal`, {
+            side: "left",
+            title: constants.MIRROR_HORIZONTAL_BUTTON,
+            icon: "mirror-horizontal",
+            onClick: this.#onHorizontalMirror.bind(this)
+        });
+
+        this.#tokenHUDManager.registerButton(`${constants.MODULE_NAME}.mirror-vertical`, {
+            side: "left",
+            title: constants.MIRROR_VERTICAL_BUTTON,
+            icon: "mirror-vertical",
+            onClick: this.#onVerticalMirror.bind(this)
+        });
+
+        this.#tokenHUDManager.registerButton(`${constants.MODULE_NAME}.toggle-afk`, {
+            side: "right",
+            title: constants.TOGGLE_AFK_BUTTON,
+            icon: "toggle-afk",
+            onClick: this.#onToggleAFK.bind(this)
         });
     }