+++ Changelog.md
@@ -1,5 +1,6 @@
 # v2.1.0
 * Added a `Shift+K` (default binding) shortcut to mark a token as AFK
+* Added token HUD buttons for mirroring tokens and marking tokens as AFK
 * Restructured the mod internals to allow for more token related features to be added
 
 # v2.0.0